Correctly handle Ctrl-c
When Ctrl-c is pressed while the finder is active, we just leave the
finder buffer's focus without closing it properly. That is annoying and
while usage of Ctrl-c is to be minimized, it happens on occasion. To
work around this problem this change adds a custom key binding causing
Ctrl-c to simply close the buffer as a <Esc> would.